requirements.txt
django~=4.1.5
python-dotenv~=0.21.1
psycopg2-binary~=2.9.5
django-qsstats-magic~=1.1.0
python-dateutil~=2.8.2
django-google-charts~=0.1.1
python-telegram-bot~=13.13.0
celery~=5.2.7
redis~=4.4.2
django-celery-beat~=2.4.0